- WebSecurityConfig: bestellingen en bestelregels verwijderd uit geblockte urls voor klanten - BestellingController & BestelregelController kunnen alle mappings nu afhandelen - domein/Bestelregel heeft nu als id een long ipv een int - domein/klant heeft nu een equals methode - bestelregels.html heeft nu ook een link naar de muteerpagina - bestelregelMuteer.html toegevoegd

+++ src/main/java/springframework/secureWeb/domein/Bestelregel.java
@@ -3,6 +3,7 @@ package springframework.secureWeb.domein;
 import java.io.Serializable;
 import java.math.BigDecimal;
 
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -22,7 +23,7 @@ public class Bestelregel implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private long id;
 	
 	@ManyToOne
     private Artikel artikel;
@@ -39,11 +40,11 @@ public class Bestelregel implements Serializable {
 	@OnDelete(action = OnDeleteAction.CASCADE)
 	private Bestelling bestelling;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
